add: add device preview and some fix

+++ lib/features/bet_history/views/adaptive_bet_history/desktop_bet_history.dart
@@ -297,7 +297,7 @@ class _DesktopBetHistoryTableHeading extends StatelessWidget {
         children: tableHeadingsWithWidth.keys
             .map(
               (entry) => SizedBox(
-                width: tableHeadingsWithWidth[entry]!.toDouble() as double,
+                width: tableHeadingsWithWidth[entry]!.toDouble(),
                 child: Text(entry, style: Styles.betHistoryDesktopField),
               ),
             )
